US_Health_Insurance_predictive_model
Our project aims to determine the extent to which various socioeconomic
and demographic parameters
influence people's decision to not opt for a health insurance in various
states within the United States. By
analyzing the mentioned problem statement, we can recognize the part of
population most deprived of
healthcare and infer the main reason for people opting not enrolling for
health-insurance. It might be helpful
to restructure policies and schemes considering some biases in the
design of health-insurance schemes.